fix(types): complete mypy coverage and fix the bugs it surfaced
Bring `mypy openfatture/` to zero errors (from 184) without type: ignore,
ignore_errors, or per-module error suppression. The two existing escape
hatches are removed: the voice module's disable_error_code list and the
"planned modules" override for payment imports.

Stubs and third-party typing:
- add lxml-stubs and types-python-dateutil
- keep ignore_missing_imports only for genuinely untyped libraries
  (sklearn, ofxparse, pyasn1_modules) and the optional nest_asyncio
- declare the lazy provider package's exports under TYPE_CHECKING so
  BaseLLMProvider and friends stop resolving to Any downstream

Bugs the type errors exposed:
- import_bank_transactions imported bank_import_service and ofx_parser,
  neither of which exists: the tool raised ImportError on every call.
  Rewritten on OFXImporter plus BankTransactionRepository, deduplicating
  by FITID against the transactions already stored for the account.
- lightning create_invoice passed `description=` to a method taking
  `descrizione` and then called _replace() on a plain dataclass, so both
  branches were broken; the amount branch now goes through the LND client.
- LightningAMLAlertEvent has no `timestamp` attribute (it is occurred_at).
- cash_flow_analysis never normalised the dict returned by graph.ainvoke(),
  unlike the other two workflows.
- the BTC/EUR rate was stored as float into a Numeric column.
- settled zero-amount invoices passed amount_msat=None into an int event
  field; the settled amount now comes from LND's amt_paid_msat, and
  simulate_payment takes an explicit amount with a documented default.
- both retry_with_backoff decorators fell out of the loop returning None
  instead of raising the last exception.
- PrivateKeyType drifted from cryptography's own PrivateKeyTypes alias,
  missing the ML-DSA/ML-KEM key types.

The single cast in fatturapa.py is documented: lxml-stubs 0.5.1 types the
nsmap argument as Mapping[str, str] while lxml accepts None as the prefix
for the default namespace.

+++ openfatture/ai/ml/models/ensemble.py
@@ -441,7 +441,7 @@ class CashFlowEnsemble:
         predictions = self.predict(X)
         y_pred = np.array([p.yhat for p in predictions])
 
-        mae = np.mean(np.abs(y.values - y_pred))
+        mae = float(np.mean(np.abs(y.values - y_pred)))
 
         logger.info("ensemble_scored", mae=mae, sample_count=len(X))
